usync: drop call to floor
'row' should never be negative, and because C/C++ defines round
towards zero for float to in casts, we should be good by simply
casting.

This change cause some problems for users of /QIfist, which is
propular amongst 4k coders. To remedy this, they should drop
QIfist, and instead include this SSE-based drop-in replacement
of ftol2:

__declspec(naked) extern "C" void _ftol2_sse()
{
	__asm {
		fstp        dword ptr[esp - 4]
		movss       xmm0, dword ptr[esp - 4]
		cvttps2dq   xmm0, xmm0
		movd        eax, xmm0
		ret
	}
}