editor: allow configuring server-options
It's useful to be able to control what network interfaces and ports
are used by the editor. So let's add support for this, and save the
configuration in the application configuration.

While we're at it, only listen locally by default, as this reduce
the firewall noise for people who only use rocket against localhost
anyway.